Generalize banks code
We want to be able to handle more banks, so now we can add that
ability by just dropping bank plugins in the banks directory.

Bank classes need to inherit from the Bank class. They should have a
name property that matches how the bank will be referred to in the
config.json file.